auto completion is now able to resolve stuff from other opened files now
there is still a problem that I wanted to remove; the entire source file being parsed **while** the lookup happens. I could not prevent this from happening
because of the way the code is written.
I moved the creation of the symbol table to the background thread; which is good.