all symbol table and resource finder tests pass
Had to add another callback to the parser observers; one that would give the position of a function
or method end so that the symbol table can determine the scope.

Also, since the native functions PHP file is now parsed with the bison parser [and all functions are
prefixed so that lint check passes) the resource finder class had to be updated to remove
the prefix before caching.