added a new PHP action to detect framework resources
this ended up being a fairly big re-write of PHP-to-editor communication; created a class hierarchy for
more code sharing. Needed to work around the external process hanging on the new action (resources) by
shoving the external process output to a file.  The new dynamic resource code is now being executed but
not yet working. I suspect it is because since the dynamic resource code runs in parallel with the
indexing code; the dynamic resources that are added get overwritten when project indexing ends.