parser now captures array keys from hard-coded arrays and array key assignments
This change required re-doing a lot of stuff with SymbolClass; SymbolClass was not adequate to hold information about the
variable being assigned and the expression that was assigned to it.  This also necessitated quite a bit of
refactoring of the ParserObserver;

Also created a new scope class to keep track of the namespace-class-method trio.