Resource parser for PHP 5.5
See http://php.net/manual/en/migration55.new-features.php
for an explanantion of the new features:

1. yield keyword
2. finally keyword
3. able to use list() in foreach()
4. able to use any expression in empty()
5. Array and String literal dereferencing: "hello"[1], ["hel", "lo"][1]

There are also a few fixes to the FullParserObserverClass so
that it outputs the same callbacks as the ResourceParserObserverClass.
Don't want to have to implemnt both the resource and the full parser
for 5.5 right now. Changes to the full parser observer are:

a.) When reporting the class name, always use "\" for the namespace
    name.
b.) Report `include` expressions located inside unary expressions
  `@include($file)`. For this, we need to recurse inside each
  top-level expression.
c.) When reporting namespaces, always use the full qualified
    namespace name (start with "\")