PHP 5.6 parser
* Constant Expressions.
* Variadic parameters in function declarations

    function add(...$numbers) {}

The parser will notify the function and method observers
when methods use variable arguments.

* Argument unpacking. `f(...$items)`
There is no AST support, but the parser will handle the ellipsis.

* POW operator  `**`

* Function and constant importing and aliasing.

    use function Name\Space\F;
    use constant Name\Space\MAX;

The parser will automatically resolve all function and constant
aliases, just like it does for unqualifed class names.

Changes to how parameters list are handled. This was needed due
to the change in the paramter list BNF.